implement suggestions
