+++ lib/constants/index.ts
@@ -1,3 +1,2 @@
 export * from './platforms';
-export type { Category } from './category';
-export { CategoryNames } from './category';
+export type * from './category';
